+++ SGGEngine/src/data/Vector3.cpp
@@ -10,11 +10,11 @@ namespace SG
 	{
 	}
 
-	Vector3::Vector3(float xy) : X(xy), Y(xy), Z(0)
+	Vector3::Vector3(int xy) : X(xy), Y(xy), Z(0)
 	{
 	}
 
-	Vector3::Vector3(float x, float y, float z) : X(x), Y(y), Z(z)
+	Vector3::Vector3(int x, int y, int z) : X(x), Y(y), Z(z)
 	{
 	}
